+++ code/core/src/main/java/org/betonquest/betonquest/quest/event/time/TimeEventFactory.java
@@ -4,7 +4,6 @@ import org.betonquest.betonquest.api.common.function.ConstantSelector;
 import org.betonquest.betonquest.api.common.function.Selector;
 import org.betonquest.betonquest.api.common.function.Selectors;
 import org.betonquest.betonquest.api.instruction.Instruction;
-import org.betonquest.betonquest.api.instruction.argument.Argument;
 import org.betonquest.betonquest.api.instruction.variable.Variable;
 import org.betonquest.betonquest.api.quest.PrimaryServerThreadData;
 import org.betonquest.betonquest.api.quest.QuestException;
@@ -25,6 +24,7 @@ import org.jetbrains.annotations.Nullable;
  * Factory to create time events from {@link Instruction}s.
  */
 public class TimeEventFactory implements PlayerEventFactory, PlayerlessEventFactory {
+
     /**
      * Server to use for fetching worlds.
      */
@@ -61,28 +61,10 @@ public class TimeEventFactory implements PlayerEventFactory, PlayerlessEventFact
     }
 
     private NullableEventAdapter createTimeEvent(final Instruction instruction) throws QuestException {
-        final String timeString = instruction.next();
-        final Time time = parseTimeType(timeString);
-        final Variable<Number> rawTime = parseTime(instruction, timeString, time != Time.SET);
+        final Variable<TimeChange> timeVariable = instruction.get(TimeParser.TIME);
         final Selector<World> worldSelector = parseWorld(instruction.getValue("world"));
         final boolean hourFormat = !instruction.hasArgument("ticks");
-        return new NullableEventAdapter(new TimeEvent(time, rawTime, worldSelector, hourFormat));
-    }
-
-    private Time parseTimeType(final String timeString) throws QuestException {
-        if (timeString.isEmpty()) {
-            throw new QuestException("Time cannot be empty");
-        }
-        return switch (timeString.charAt(0)) {
-            case '+' -> Time.ADD;
-            case '-' -> Time.SUBTRACT;
-            default -> Time.SET;
-        };
-    }
-
-    private Variable<Number> parseTime(final Instruction instruction, final String timeString, final boolean cutFirst) throws QuestException {
-        final String rawTime = cutFirst ? timeString.substring(1) : timeString;
-        return instruction.get(rawTime, Argument.NUMBER);
+        return new NullableEventAdapter(new TimeEvent(timeVariable, worldSelector, hourFormat));
     }
 
     private Selector<World> parseWorld(@Nullable final String worldName) {